+++ src/importer/exmaralda/tests.rs
@@ -45,6 +45,66 @@ fn test_exb_fail_for_timeline() {
 }
 
 #[test]
+fn test_exb_fail_for_no_category() {
+    let import = ImportEXMARaLDA::default();
+    let import_path = "./tests/data/import/exmaralda/fail-no_category/";
+    let (sender, receiver) = mpsc::channel();
+    let r = import.import_corpus(Path::new(import_path), &BTreeMap::new(), Some(sender));
+    assert!(r.is_ok());
+    assert!(receiver.into_iter().count() > 0);
+    let document_path = Path::new(import_path).join("test_doc.exb");
+    let mut u = GraphUpdate::default();
+    assert!(import
+        .import_document(
+            Path::new(import_path),
+            document_path.as_path(),
+            &mut u,
+            &None
+        )
+        .is_err());
+}
+
+#[test]
+fn test_exb_fail_for_no_speaker() {
+    let import = ImportEXMARaLDA::default();
+    let import_path = "./tests/data/import/exmaralda/fail-no_speaker/";
+    let (sender, receiver) = mpsc::channel();
+    let r = import.import_corpus(Path::new(import_path), &BTreeMap::new(), Some(sender));
+    assert!(r.is_ok());
+    assert!(receiver.into_iter().count() > 0);
+    let document_path = Path::new(import_path).join("test_doc.exb");
+    let mut u = GraphUpdate::default();
+    assert!(import
+        .import_document(
+            Path::new(import_path),
+            document_path.as_path(),
+            &mut u,
+            &None
+        )
+        .is_err());
+}
+
+#[test]
+fn test_exb_fail_for_undefined_speaker() {
+    let import = ImportEXMARaLDA::default();
+    let import_path = "./tests/data/import/exmaralda/fail-undefined_speaker/";
+    let (sender, receiver) = mpsc::channel();
+    let r = import.import_corpus(Path::new(import_path), &BTreeMap::new(), Some(sender));
+    assert!(r.is_ok());
+    assert!(receiver.into_iter().count() > 0);
+    let document_path = Path::new(import_path).join("test_doc.exb");
+    let mut u = GraphUpdate::default();
+    assert!(import
+        .import_document(
+            Path::new(import_path),
+            document_path.as_path(),
+            &mut u,
+            &None
+        )
+        .is_err());
+}
+
+#[test]
 fn test_fail_invalid() {
     let import = ImportEXMARaLDA::default();
     let import_path = "./tests/data/import/exmaralda/fail-invalid/import/";